Avoid checking assert with unsigned integer cast
This fixes #209.

+++ Source/FamiTrackerView.cpp
@@ -1778,21 +1778,21 @@ void CFamiTrackerView::MoveCursorPrevChannel()
 
 void CFamiTrackerView::SelectFrame(unsigned int Frame)
 {
-	ASSERT(Frame < MAX_FRAMES);
+	ASSERT(static_cast<int>(Frame) < MAX_FRAMES);		// avoid comparing against UINT32_MAX
 	m_pPatternEditor->MoveToFrame(Frame);
 	InvalidateCursor();
 }
 
 void CFamiTrackerView::SelectRow(unsigned int Row)
 {
-	ASSERT(Row < MAX_PATTERN_LENGTH);
+	ASSERT(static_cast<int>(Row) < MAX_PATTERN_LENGTH);		// avoid comparing against UINT32_MAX
 	m_pPatternEditor->MoveToRow(Row);
 	InvalidateCursor();
 }
 
 void CFamiTrackerView::SelectChannel(unsigned int Channel)
 {
-	ASSERT(Channel < MAX_CHANNELS);
+	ASSERT(static_cast<int>(Channel) < MAX_CHANNELS);		// avoid comparing against UINT32_MAX
 	m_pPatternEditor->MoveToChannel(Channel);
 	InvalidateCursor();
 }
